Add orders and pickup points, enhance product admin features and UI

+++ app/repositories/product_repository.py
@@ -1,7 +1,8 @@
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
-from app.exceptions.exceptions import DataNotFoundError
+from app.exceptions.exceptions import DataNotFoundError, InvalidEnteredDataError
+from app.models import OrderItem
 from app.models.product_model import Product
 
 
@@ -47,6 +48,11 @@ class ProductRepository:
 
         if product is None:
             raise DataNotFoundError()
+
+        has_order_items_stmt = select(OrderItem).where(OrderItem.product_id == product_id).limit(1)
+        has_order_items = self.db.execute(has_order_items_stmt).scalars().first()
+        if has_order_items is not None:
+            raise InvalidEnteredDataError()
 
         self.db.delete(product)
         self.db.commit()
